requirements.txt
bcrypt==3.2.0
cffi==1.15.0
colorama==0.4.4
cryptography==36.0.1
greenlet==1.1.2
paramiko==2.9.2
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.31
